backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
transformers>=4.40.0
sentence-transformers>=2.6.1
torch
